# swagger.yaml
openapi: 3.0.0
info:
  title: July Project Backend API
  version: 1.1.0
  description: API documentation for authentication, users, and artworks modules.
servers:
  - url: http://localhost:3001/api/v1
...
paths:
...
  /artworks:
    get:
      tags: [Artworks]
      summary: List artworks
      parameters:
        - $ref: '#/components/parameters/CategoryQuery'
        - $ref: '#/components/parameters/ArtistIdQuery'
      responses:
        "200":
          description: Artworks fetched
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  message:
                    type: string
                    example: Artworks retrieved successfully
                  data:
                    type: object
                    properties:
                      artworks:
                        type: array
                        items:
                          $ref: '#/components/schemas/ArtworkResponse'

    post:
      tags: [Artworks]
      summary: Create artwork
      description: Requires artist role. Uses multipart/form-data with image files.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/CreateArtworkFormData'
      responses:
        "201":
          description: Artwork created
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  message:
                    type: string
                    example: Artwork created successfully
                  data:
                    type: object
                    properties:
                      artwork:
                        $ref: '#/components/schemas/ArtworkResponse'
        "400":
          $ref: '#/components/responses/ValidationError'
        "401":
          $ref: '#/components/responses/UnauthorizedError'
        "403":
          $ref: '#/components/responses/ForbiddenError'
...
  /artists:
    get:
      tags: [Artists]
      summary: List all artists
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Artists fetched successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ArtistSummary'
        "401":
          $ref: '#/components/responses/UnauthorizedError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    CategoryQuery:
      name: category
      in: query
      required: false
      schema:
        type: string
        enum:
          - لوحات فنية
          - تطريز فلسطيني
          - خزف وفخار
          - خط عربي
          - تصوير فوتوغرافي
          - نحت ومجسمات
    ArtistIdQuery:
      name: artist_id
      in: query
      required: false
      schema:
        type: string
        format: uuid
  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    UnauthorizedError:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ForbiddenError:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        status:
          type: string
          example: error
        message:
          type: string
          example: Something went wrong
...
    ArtworkImage:
      type: object
      properties:
        id:
          type: string
          format: uuid
        filename:
          type: string
        alt_text:
          type: string
          nullable: true
        is_featured:
          type: boolean
        sort_order:
          type: integer
    ArtworkArtistInfo:
      type: object
      properties:
        artist_name:
          type: string
          nullable: true
        first_name:
          type: string
          nullable: true
        last_name:
          type: string
          nullable: true
        profile_image:
          type: string
          format: uri
          nullable: true
        location:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
    ArtworkResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        category:
          type: string
          enum:
            - لوحات فنية
            - تطريز فلسطيني
            - خزف وفخار
            - خط عربي
            - تصوير فوتوغرافي
            - نحت ومجسمات
        artist_id:
          type: string
          format: uuid
        price:
          type: number
        quantity:
          type: integer
        is_active:
          type: boolean
        deleted_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        users:
          oneOf:
            - $ref: '#/components/schemas/ArtworkArtistInfo'
            - type: array
              items:
                $ref: '#/components/schemas/ArtworkArtistInfo'
        artwork_images:
          type: array
          items:
            $ref: '#/components/schemas/ArtworkImage'
    CreateArtworkFormData:
      type: object
      required: [title, description, category, price, quantity, images]
      properties:
        title:
          type: string
          minLength: 2
        description:
          type: string
          minLength: 10
        category:
          type: string
          enum:
            - لوحات فنية
            - تطريز فلسطيني
            - خزف وفخار
            - خط عربي
            - تصوير فوتوغرافي
            - نحت ومجسمات
        price:
          type: number
          minimum: 0
        quantity:
          type: integer
          minimum: 0
        mainImageIndex:
          type: integer
          minimum: 0
          example: 0
        images:
          type: array
          maxItems: 3
          items:
            type: string
            format: binary
...
    ArtistSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        first_name:
          type: string
        last_name:
          type: string
        artist_name:
          type: string
          nullable: true
        artist_since:
          type: string
          format: date-time
          nullable: true
        profile_image:
          type: string
          format: uri
          nullable: true
        bio:
          type: string
          nullable: true
        location:
          type: string
          nullable: true
        role:
          type: string
          example: artist
...